prompt-manager/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic
python-dotenv
pyyaml
dash
